Criado a entidade empresa

+++ Domain/Models/Vagas.cs
@@ -27,7 +27,7 @@ namespace JobApplication.Domain.Models
         [Required]
         [Column("salario")]
         [MaxLength(10)]
-        public double salario { get; set; }
+        public decimal salario { get; set; }
 
         [Required]
         [Column("localizacao")]
